requirements.txt
astroid==2.0.4
colorama==0.4.0
cycler==0.10.0
isort==4.3.4
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
matplotlib==3.0.2
mccabe==0.6.1
numpy==1.15.4
opencv-python==3.4.3.18
Pillow==5.3.0
pylint==2.1.1
pyparsing==2.3.0
python-dateutil==2.7.5
six==1.11.0
typed-ast==1.1.0
wrapt==1.10.11
